refactor: ODataExpression for calling build method
