+++ src/lib/ODataExpression.ts
@@ -1,18 +1,21 @@
-import type { Expression } from "./Expression";
 import type { ExcludeProperties } from "./ExcludeProperties";
 import { ODataQueryBase } from "./ODataQueryBase";
 import { ODataV4QueryProvider } from "./ODataV4QueryProvider";
 
 /**
- * Represents a query expression builder only, without any provider.
+ * Represents a query expression builder only, with empty endpoiont.
  */
 
-export class ODataExpression<T, U = ExcludeProperties<T, unknown[]>>  extends ODataQueryBase<T, U>  {
+export class ODataExpression<T, U = ExcludeProperties<T, unknown[]>>  {
+
+  oDataQuery: ODataQueryBase<T, U>;
+
   static forV4<T>() {
-    return new ODataExpression<T>();
+    return new ODataExpression<T>().oDataQuery;
   }
 
-  constructor(public readonly expression?: Expression) {
-    super(new ODataV4QueryProvider(''), expression);
+  constructor() {
+    this.oDataQuery = new ODataQueryBase<T, U>(new ODataV4QueryProvider(''));
   }
+  
 }
